[追加]電波に当たった時に電気のパーティクルが発生するように設定
